requirements.txt
django==4.2.0
djangorestframework==3.14.0
pyjwt==2.6.0
celery==5.2.7
redis==4.5.1
mssql-django==1.3
pyodbc==4.0.35